add bigint

+++ kubeojo/priv/repo/migrations/20180716205222_create_tests_failures.exs
@@ -4,10 +4,10 @@ defmodule Kubeojo.Repo.Migrations.CreateTestsFailures do
   def change do
     create table(:tests_failures) do
        add :testname, :string
-       add :count_failed, :integer
-       add :build_timestamp, :integer
+       add :count_failed, :bigint
+       add :build_timestamp, :bigint
        add :jobname, :string
-       add :jobnumber, :integer
+       add :jobnumber, :bigint
 
        timestamps()
     end
